Update Interop with newer FFL, change library name to "ffl"
When running the CMake build for FFL standalone, it's "ffl.dll/lib" on
Windows by default and "libffl.so/a" on Linuxy systems, and FFLSharp had
been using the libffl name. Changing it to "ffl" here doesn't seem to
make a difference as I'm sure .NET will know to look for "libffl" if I
specify "ffl", so, yes.

The version of FFL here should... be up to date with my fork by the time
I finally release the texture callback and, by extension, this.